+++ Pfhorge Source/View and Controller/Others/PhColorListController.m
@@ -317,9 +317,9 @@
 // *********************** Table Updater Methods ***********************
 
 - (void)tableView:(NSTableView *)view
-    willDisplayCell:(id)cell
-    forTableColumn:(NSTableColumn *)col
-    row:(int)row
+  willDisplayCell:(id)cell
+   forTableColumn:(NSTableColumn *)col
+			  row:(NSInteger)row
 {
     if (row == 0)
         [cell setBackgroundColor:[NSColor whiteColor]];
@@ -329,14 +329,14 @@
     
 }
 
-- (int)numberOfRowsInTableView:(NSTableView *)view
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)view
 {
     return [names count] + 1;
 }
 
 - (id)tableView:(NSTableView *)view
-    objectValueForTableColumn:(NSTableColumn *)col
-    row:(int)row
+objectValueForTableColumn:(NSTableColumn *)col
+			row:(NSInteger)row
 {
     if (row == 0)
         return @"Double Click For New";
@@ -345,8 +345,8 @@
 }
 
 - (BOOL)tableView:(NSTableView *)aTableView
-    shouldEditTableColumn:(NSTableColumn *)col
-    row:(int)rowIndex
+shouldEditTableColumn:(NSTableColumn *)col
+			  row:(NSInteger)rowIndex
 {
     NSParameterAssert(rowIndex >= 0 && rowIndex < (((int)[names count]) + 1));
     //NSLog(@"editing somthing...");
@@ -421,9 +421,9 @@
 }
 
 - (void)tableView:(NSTableView *)aTableView
-    setObjectValue:anObject
-    forTableColumn:(NSTableColumn *)col
-    row:(int)rowIndex
+   setObjectValue:anObject
+   forTableColumn:(NSTableColumn *)col
+			  row:(NSInteger)rowIndex
 {
     // Should never get here for right now, acutally...
     SEND_ERROR_MSG(@"Color List Table Attempted To Set Object.");
